+++ packages/terra-action-header/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Changed
+  * Updated jest snapshots for icon button changes.
+
 ## 2.75.0 - (December 7, 2022)
 
 * Changed